# Add cache for methods