The PR adds cache for methods which will speed up all the code going through `faker#expression`

I made some measurements

before
```
Benchmark                            Mode  Cnt    Score   Error   Units
JmhTest._bothifyExpression          thrpt    5  157.984 ± 2.141  ops/ms
JmhTest._letterifyExpression        thrpt    5  161.506 ± 1.238  ops/ms
JmhTest._optionsExpression          thrpt    5  126.751 ± 6.211  ops/ms
JmhTest._regexifyExpression         thrpt    5  120.731 ± 2.774  ops/ms
JmhTest.oneMethodCallExpression     thrpt    5  124.431 ± 0.641  ops/ms
JmhTest.threeMethodsCallExpression  thrpt    5   31.784 ± 0.090  ops/ms

```

after
```
Benchmark                            Mode  Cnt    Score    Error   Units
JmhTest._bothifyExpression          thrpt    5  533.255 ± 22.638  ops/ms
JmhTest._letterifyExpression        thrpt    5  531.161 ± 12.172  ops/ms
JmhTest._optionsExpression          thrpt    5  370.678 ± 16.982  ops/ms
JmhTest._regexifyExpression         thrpt    5  238.741 ± 15.081  ops/ms
JmhTest.oneMethodCallExpression     thrpt    5  337.323 ±  1.139  ops/ms
JmhTest.threeMethodsCallExpression  thrpt    5   67.555 ±  6.570  ops/ms
```
So, it show that it this kind of caching speeds up from 2 to 3 times for different cases

and code for benchmark
```java


@BenchmarkMode(Mode.AverageTime)
@OutputTimeUnit(TimeUnit.MILLISECONDS)
@State(Scope.Benchmark)
@Fork(value = 2, jvmArgs = {"-Xms2G", "-Xmx2G"})
public class JmhTest {

    private static final Faker FAKER = new Faker();
    public static void main(String[] args) throws RunnerException {

        Options opt = new OptionsBuilder()
                .include(JmhTest.class.getSimpleName())
                .forks(1)
                .build();

        new Runner(opt).run();
    }

    @Benchmark
    @BenchmarkMode(Mode.Throughput)
    public void _optionsExpression(Blackhole blackhole) {
        blackhole.consume(FAKER.expression("#{options.option 'a','b','c','d'}"));
    }

    @Benchmark
    @BenchmarkMode(Mode.Throughput)
    public void _letterifyExpression(Blackhole blackhole) {
        blackhole.consume(FAKER.expression("#{letterify '????','true'}"));
    }

    @Benchmark
    @BenchmarkMode(Mode.Throughput)
    public void _regexifyExpression(Blackhole blackhole) {
        blackhole.consume(FAKER.expression("#{regexify '\\.\\*\\?\\+'}"));
    }

    @Benchmark
    @BenchmarkMode(Mode.Throughput)
    public void _bothifyExpression(Blackhole blackhole) {
        blackhole.consume(FAKER.expression("#{bothify '????','false'}"));
    }

    @Benchmark
    @BenchmarkMode(Mode.Throughput)
    public void oneMethodCallExpression(Blackhole blackhole) {
        blackhole.consume(FAKER.expression("#{number.number_between '1','10'}"));
    }

    @Benchmark
    @BenchmarkMode(Mode.Throughput)
    public void threeMethodsCallExpression(Blackhole blackhole) {
        blackhole.consume(FAKER.expression("#{Name.first_name} #{Name.first_name} #{Name.last_name}"));
    }
}
```