@bodiam 2022-02-11
Awesome! What would you think about adding the performance tests to the repo too?