@snuyanzin 2022-02-11
As far as I know there is kind of weird license in case it is published... Some projects do it in a separate module. I think this way would be safer