@bodiam 2022-02-11
Ah, it's the GPL license. That's always a bit tricky indeed. Okay, makes sense!